+++ HISTORY.rst
@@ -2,6 +2,16 @@
 History
 =======
 
+2.3 (unreleased)
+----------------
+
+- Use ``yafowil.utils.entry_point`` decorator.
+  [rnix]
+
+- Plone 5 support.
+  [rnix]
+
+
 2.2 (2015-10-10)
 ----------------